@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import App from './App';
 import {Provider} from 'react-redux'
 import { createStore } from 'redux'
@@ -7,10 +6,15 @@ import { reducer } from './reducer';
 
 const store = createStore(reducer)
 
-ReactDOM.render(
+import { createRoot } from 'react-dom/client';
+
+window.onload = () => {
+
+  const root = createRoot(document.getElementById('root')!);
+  
+  root.render(
     <Provider store={store}>
       <App/>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 );
-
+}
