@@ -1,10 +1,6 @@
 import yargs from "yargs";
-import { checkElementContainsText, checkElementExists, clickElement, startBrowser } from "./playwrightAccess";
-import { config, FrameworkData, initializeFrameworks, BenchmarkOptions } from "./common";
-
-import * as R from "ramda";
-import { valid } from "semver";
-import { ElementHandle, Page } from "playwright";
+import { checkElementContainsText, checkElementExists, clickElement, startBrowser } from "./playwrightAccess.js";
+import { config, FrameworkData, initializeFrameworks, BenchmarkOptions } from "./common.js";
 
 let args: any = yargs(process.argv)
   .usage("$0 [--framework Framework1 Framework2 ...] [--benchmark Benchmark1 Benchmark2 ...]")
@@ -114,7 +110,7 @@ let benchmarkOptions: BenchmarkOptions = {
 };
 async function main() {
   if (args.help) {
-    yargs.showHelp();
+    // yargs.showHelp();
   } else {
     await runBench(runFrameworks);
   }
